feat: place buildings and rotate
Building prefab objects can now be placed on a grid.
These buildings can also be rotated and no two images can be placed on the same space.

however the system currently has a bug, whereby the buildings cannot be placed on the spot of the mouse
issue #7 for more information

resolves #2